+++ source/ProcessManager.Client.Tests/ProcessManager.Client.Tests.csproj
@@ -8,14 +8,25 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Include="..\ProcessManager.Core.Tests\Fixtures\ProcessManagerDatabaseManager.cs" Link="Fixtures\ProcessManagerDatabaseManager.cs" />
+    <Compile Include="..\ProcessManager.Orchestrations.Tests\Fixtures\OrchestrationsAppFixtureBase.cs" Link="Fixtures\OrchestrationsAppFixtureBase.cs" />
+    <Compile Include="..\ProcessManager.Tests\Fixtures\ProcessManagerAppFixtureBase.cs" Link="Fixtures\ProcessManagerAppFixtureBase.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="coverlet.collector" Version="6.0.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
     <PackageReference Include="xunit" Version="2.9.2" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.5.3" />
+    <PackageReference Include="Energinet.DataHub.Core.FunctionApp.TestCommon" Version="7.0.0" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\ProcessManager.Client\ProcessManager.Client.csproj" />
+    <ProjectReference Include="..\ProcessManager.Core\ProcessManager.Core.csproj" />
+    <ProjectReference Include="..\ProcessManager.DatabaseMigration\ProcessManager.DatabaseMigration.csproj" />
+    <ProjectReference Include="..\ProcessManager.Orchestrations\ProcessManager.Orchestrations.csproj" />
+    <ProjectReference Include="..\ProcessManager\ProcessManager.csproj" />
   </ItemGroup>
 
   <ItemGroup>
@@ -23,8 +34,16 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Fixtures\" />
     <Folder Include="Unit\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="functionapphost.settings.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Update="integrationtest.local.settings.json">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 
 </Project>
