+++ source/ProcessManager.Client.Tests/integrationtest.local.settings.sample.json
@@ -1,0 +1,3 @@
+﻿{
+  "AZURE_KEYVAULT_URL": "<Integration test key vault url>"
+}
